adds logging for parser and bridge metadata (#2192)
* adds logging for parser and bridge metadata

As part of adding more logging we refactor the logging of passes into
a Logging::PassManager, which is essentially a wrapper for the main
PassManager that overrides the init_apply and end_apply methors to
redirect logging. We replace the previous custom init_apply/end_apply
with inheritance from Logging::PassManager.

Of course, it can't be as simple, since parser logging is spread
through pass managers, inspectors, and functions. So we have added
an option to the FileLog object to append to an existing log.
With this support, we dump all parser logs in one file.

We publish the logs in the manifest file. With that come a set of
other changes, in particular related to handling run_id. Because it
uses the openssl library to compute the SHA, and thus ends up included
in gtests, we factor out the constructor and include the ssl library
only in a .cpp file. Now manifest.h can be safely included anywhere.

Finally, we make the `--verbose` option to the driver compatible with
Glass and take a verbosity argument. This is used mainly for PHV and
parser logging. Verbose 1 is LOG1, 2 is LOG3 or LOG 4 (depending on
module), and 3 is truly verbose -- 250MB of log file for a switch
compilation. And for good measure, we generate two of these, one for
each allocation pass. The more passes, the more logs.

Multiple fixes revealed by adding tests:
- add missing includes
- logs should not be generated unless verbose is > 0
- logs file path in the manifest is relative to the output directory
- add at least one verbose test to the option testing suite
- update the options in the test script (archive with name and verbose)

+++ bf-asm/counter.cpp
@@ -89,7 +89,7 @@ void CounterTable::pass1() {
         else if ((l.threshold & 0xf) != 0)
             error(l.lineno, "lrt threshold must be a mulitple of 16");
         if (type & BYTES) {
-            if (l.interval > MAX_LRT_BYTE_INTERVAL)
+            if (static_cast<uint64_t>(l.interval) > MAX_LRT_BYTE_INTERVAL)
                 error(l.lineno, "lrt interval too large");
             else if ((l.interval & 0xff) != 0)
                 error(l.lineno, "lrt interval must be a mulitple of 256 when measuring bytes");
@@ -151,12 +151,12 @@ template<class REGS> void CounterTable::write_merge_regs(REGS &regs, MatchTable 
     } else if (args[0].type == Table::Call::Arg::Field) {
         auto addr = args[0].field();
         auto address_bits = addr->size;
-        adr_mask |= ((1U << address_bits) - 1) << (counter_shifts[format->groups()]); 
+        adr_mask |= ((1U << address_bits) - 1) << (counter_shifts[format->groups()]);
     }
 
-    
+
     if (args[1].type == Table::Call::Arg::Name && strcmp(args[1].name(), "$DEFAULT") == 0) {
-        adr_default = (1U << STATISTICS_PER_FLOW_ENABLE_START_BIT); 
+        adr_default = (1U << STATISTICS_PER_FLOW_ENABLE_START_BIT);
     } else if (args[1].type == Table::Call::Arg::Field) {
         if (args[0].type == Table::Call::Arg::Field) {
             per_entry_en_mux_ctl = args[1].field()->bit(0) - args[0].field()->bit(0) ;
